+++ src/datatypes/shapearray.ts
@@ -1,14 +1,17 @@
 import type {EntityId} from '../entity';
 import type {Buffers} from '../buffers';
 import type {ComponentType} from '../component';
+import type {ComponentEnum} from '../enums';
+import {InternalError} from '../errors';
 
 export interface ShapeArray {
   syncThreads(): void;
   set(entityId: EntityId, type: ComponentType<any>): void;
   unset(entityId: EntityId, type: ComponentType<any>): void;
   isSet(entityId: EntityId, type: ComponentType<any>): boolean;
+  get(entityId: number, enumeration: ComponentEnum): number;
   clear(): void;
-  match(entityId: EntityId, positiveMask: number[]): boolean;
+  match(entityId: EntityId, positiveMask: number[], positiveValues: number[]): boolean;
   matchNot(entityId: EntityId, negativeMask: number[]): boolean;
 }
 
@@ -17,8 +20,8 @@ export class UnsharedShapeArray implements ShapeArray {
   private readonly stride: number;
   private array: Uint32Array;
 
-  constructor(bufferKey: string, numComponentTypes: number, maxEntities: number, buffers: Buffers) {
-    this.stride = Math.ceil(numComponentTypes / 32);
+  constructor(bufferKey: string, numBits: number, maxEntities: number, buffers: Buffers) {
+    this.stride = Math.ceil(numBits / 32);
     buffers.register(
       bufferKey, maxEntities * this.stride, Uint32Array, shapes => {this.array = shapes;});
   }
@@ -31,7 +34,9 @@ export class UnsharedShapeArray implements ShapeArray {
     const binding = type.__binding!;
     const index = entityId * this.stride + binding.shapeOffset;
     const mask = binding.shapeMask;
-    this.array[index] |= mask;
+    const value = binding.shapeValue;
+    this.array[index] &= ~mask;
+    this.array[index] |= value;
   }
 
   unset(entityId: number, type: ComponentType<any>): void {
@@ -45,19 +50,30 @@ export class UnsharedShapeArray implements ShapeArray {
     const binding = type.__binding!;
     const index = entityId * this.stride + binding.shapeOffset;
     const mask = binding.shapeMask;
-    return (this.array[index] & mask) !== 0;
+    const value = binding.shapeValue;
+    return (this.array[index] & mask) === value;
+  }
+
+  get(entityId: number, enumeration: ComponentEnum): number {
+    const binding = enumeration.__binding;
+    const index = entityId * this.stride + binding.shapeOffset;
+    const mask = binding.shapeMask;
+    return (this.array[index] & mask) >>> binding.shapeShift;
   }
 
   clear(): void {
     this.array.fill(0);
   }
 
-  match(entityId: EntityId, positiveMask: number[]): boolean {
+  match(entityId: EntityId, positiveMask: number[], positiveValues: number[]): boolean {
+    DEBUG: if (positiveMask.length !== positiveValues.length) {
+      throw new InternalError(
+        `Mismatched mask and value lengths: ${positiveMask.length} vs ${positiveValues.length}`);
+    }
     const array = this.array;
     const index = entityId * this.stride;
     for (let i = 0; i < positiveMask.length; i++) {
-      const maskByte = positiveMask[i];
-      if ((array[index + i] & maskByte) !== maskByte) return false;
+      if ((array[index + i] & positiveMask[i]) !== positiveValues[i]) return false;
     }
     return true;
   }
@@ -66,8 +82,7 @@ export class UnsharedShapeArray implements ShapeArray {
     const array = this.array;
     const index = entityId * this.stride;
     for (let i = 0; i < negativeMask.length; i++) {
-      const maskByte = negativeMask[i];
-      if ((array[index + i] & maskByte) !== 0) return false;
+      if ((array[index + i] & negativeMask[i]) !== 0) return false;
     }
     return true;
   }
@@ -78,8 +93,8 @@ export class AtomicSharedShapeArray implements ShapeArray {
   private readonly stride: number;
   private array: Uint32Array;
 
-  constructor(bufferKey: string, numComponentTypes: number, maxEntities: number, buffers: Buffers) {
-    this.stride = Math.ceil(numComponentTypes / 32);
+  constructor(bufferKey: string, numBits: number, maxEntities: number, buffers: Buffers) {
+    this.stride = Math.ceil(numBits / 32);
     buffers.register(
       bufferKey, maxEntities * this.stride, Uint32Array, shapes => {this.array = shapes;});
   }
@@ -93,7 +108,9 @@ export class AtomicSharedShapeArray implements ShapeArray {
     const binding = type.__binding!;
     const index = entityId * this.stride + binding.shapeOffset;
     const mask = binding.shapeMask;
-    Atomics.or(this.array, index, mask);
+    const value = binding.shapeValue;
+    if (mask !== value) Atomics.and(this.array, index, ~mask);
+    Atomics.or(this.array, index, value);
   }
 
   unset(entityId: number, type: ComponentType<any>): void {
@@ -107,21 +124,32 @@ export class AtomicSharedShapeArray implements ShapeArray {
     const binding = type.__binding!;
     const index = entityId * this.stride + binding.shapeOffset;
     const mask = binding.shapeMask;
+    const value = binding.shapeValue;
     // Entity liveness flag can be written at any time from any thread, so do atomic check.
-    if (type.id === 0) return (Atomics.load(this.array, index) & mask) !== 0;
-    return (this.array[index] & mask) !== 0;
+    if (type.id === 0) return (Atomics.load(this.array, index) & mask) === value;
+    return (this.array[index] & mask) === value;
+  }
+
+  get(entityId: number, enumeration: ComponentEnum): number {
+    const binding = enumeration.__binding;
+    const index = entityId * this.stride + binding.shapeOffset;
+    const mask = binding.shapeMask;
+    return (this.array[index] & mask) >>> binding.shapeShift;
   }
 
   clear(): void {
     this.array.fill(0);
   }
 
-  match(entityId: EntityId, positiveMask: number[]): boolean {
+  match(entityId: EntityId, positiveMask: number[], positiveValues: number[]): boolean {
+    DEBUG: if (positiveMask.length !== positiveValues.length) {
+      throw new InternalError(
+        `Mismatched mask and value lengths: ${positiveMask.length} vs ${positiveValues.length}`);
+    }
     const array = this.array;
     const index = entityId * this.stride;
     for (let i = 0; i < positiveMask.length; i++) {
-      const maskByte = positiveMask[i];
-      if ((array[index + i] & maskByte) !== maskByte) return false;
+      if ((array[index + i] & positiveMask[i]) !== positiveValues[i]) return false;
     }
     return true;
   }
@@ -130,8 +158,7 @@ export class AtomicSharedShapeArray implements ShapeArray {
     const array = this.array;
     const index = entityId * this.stride;
     for (let i = 0; i < negativeMask.length; i++) {
-      const maskByte = negativeMask[i];
-      if ((array[index + i] & maskByte) !== 0) return false;
+      if ((array[index + i] & negativeMask[i]) !== 0) return false;
     }
     return true;
   }