+++ src/index.ts
@@ -9,3 +9,4 @@ export type {ComponentType} from './component';
 export type {Frame, SystemGroup} from './schedule';
 export {co, CanceledError} from './coroutines';
 export type {Coroutine, CurrentCoroutine} from './coroutines';
+export type {ComponentEnum} from './enums';